rf: refactor navbar-page to pages, delete hooks folder, rename product-details
