+++ src/Components/ManAccessories/man-accessories.js.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from 'react-router-dom'
+import IconName from "../Icon";
 
 const MenAccessories = ({ menData }) => {
   const { name, price } = menData || [];
@@ -8,7 +9,9 @@ const MenAccessories = ({ menData }) => {
       <div>
         <p>{name}</p>
         <p>{price}</p>
+        <div>
         <Link to={`/product-details/${name}`}>View product</Link>
+        </div>
       </div>
     </React.StrictMode>
   );
