+++ src/Pages/Home/Navbar.js
@@ -1,9 +1,9 @@
 import { NavLink } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 
-import classes from './nav.module.css'
+import classes from "./nav.module.css";
 import { person } from "react-icons-kit/iconic/person";
-import IconName from '../../Components/Icon'
+import IconName from "../../Components/Icon";
 
 export default function ({ onShowAuthModal, onOpen, showAuthModal, userName }) {
   const { t, i18n } = useTranslation();
@@ -83,8 +83,7 @@ export default function ({ onShowAuthModal, onOpen, showAuthModal, userName }) {
             <a href="#/">menu</a>
           </div>
           {/* Cart Icon */}
-          <div className="navbar-button md:hidden lg:block">
-          </div>
+          <div className="navbar-button md:hidden lg:block"></div>
           {/* Auth Page Modal nav */}
           <div>
             <button onClick={onShowAuthModal} style={{ color: "white" }}>
@@ -116,4 +115,4 @@ export default function ({ onShowAuthModal, onOpen, showAuthModal, userName }) {
       </div>
     </>
   );
-};
+}
