+++ src/index.js
@@ -6,7 +6,9 @@ import { Provider } from "react-redux";
 import { QueryClientProvider, QueryClient } from "react-query";
 import store from "./Store/store";
 import { BrowserRouter } from "react-router-dom";
-import './i18n'
+import "./i18n";
+import UseAnimations from "react-useanimations";
+import loading from "react-useanimations/lib/loading";
 
 const container = document.getElementById("root");
 const root = createRoot(container);
@@ -15,7 +17,13 @@ root.render(
     <Provider store={store}>
       <BrowserRouter>
         <QueryClientProvider client={new QueryClient()}>
-          <Suspense fallback={<div>Loading ~~~</div>}>
+          <Suspense
+            fallback={
+              <div>
+                <UseAnimations animation={loading} />
+              </div>
+            }
+          >
             <App />
           </Suspense>
         </QueryClientProvider>
